% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_calc.R
\name{age_calc}
\alias{age_calc}
\title{unction to calculate age from date of birth.}
\source{
This function was developed in part from this response on the R-Help mailing list.
\url{https://stat.ethz.ch/pipermail/r-help/2009-November/217649.html}
}
\usage{
age_calc(dob, enddate = Sys.Date(), units = "months", precise = TRUE)
}
\arguments{
\item{dob}{a vector of class \code{Date} representing the date of birth/start date}

\item{enddate}{a vector of class Date representing the when the observation's 
age is of interest, defaults to current date.}

\item{units}{character, which units of age should be calculated? allowed values are 
days, months, and years}

\item{precise}{logical indicating whether or not to calculate with leap year 
and leap second precision}
}
\value{
A numeric vector of ages the same length as the dob vector
}
\description{
his function calculates age in days, months, or years from a 
date of birth to another arbitrary date. This returns a numeric vector in 
the specified units.
}
\examples{
a <- as.Date(seq(as.POSIXct('1987-05-29 018:07:00'), len=26, by="21 day"))
b <- as.Date(seq(as.POSIXct('2002-05-29 018:07:00'), len=26, by="21 day"))

age <- age_calc(a, units='years')
age
age <- age_calc(a, units='months')
age
age <- age_calc(a, as.Date('2005-09-01'))
age
}
\author{
Jason P. Becker
}
\seealso{
See also \code{\link{difftime}} which this function uses and mimics 
some functionality but at higher unit levels.
}

