\name{eeptools-package}
\alias{eeptools-package}
\alias{eeptools}
\docType{package}
\title{
Convenience functions
}
\description{
Make R tutorials easier
}
\details{
\tabular{ll}{
Package: \tab eeptools\cr
Type: \tab Package\cr
Version: \tab 0.1 \cr
Date: \tab 2012-07-21\cr
License: \tab Public Domain \cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Jared E Knowles

Jared E. Knowles <jknowles@gmail.com>
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
