% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/proficiency_tools.R
\name{mosaictabs.label}
\alias{mosaictabs.label}
\title{Draw a visual crosstab (mosaid plot) with shading for correlations
and labels in each cell.}
\usage{
mosaictabs.label(data, rowvar, colvar, varnames, title, subtitle)
}
\arguments{
\item{data}{a data object, matrix or dataframe, that contains the categorical
variables to compose the crosstab}

\item{rowvar}{a reference to the column in data that will be displayed on the rows
of the crosstab, expressed as data$rowvar}

\item{colvar}{a reference to the column in data that will be displayed in columns of the
crosstab}

\item{varnames}{a character vector of length two with the labels for rowvar and colvar
respectively}

\item{title}{a character vector of length one that contains the main title for the plot}

\item{subtitle}{a character vector of length one that contains the subtitle displayed
beneath the plot}
}
\value{
A mosaic plot
}
\description{
Improves labeling of mosaic plots over \code{mosaic} from the vcd package
}
\examples{
df<-data.frame(cbind(x=seq(1,3,by=1), y=sample(LETTERS[6:8],60,replace=TRUE)),
fac=sample(LETTERS[1:4], 60, replace=TRUE))
varnames<-c('Quality','Grade')
mosaictabs.label(df,df$y,df$fac,varnames,'My Plot','Foo')
}
\seealso{
\code{\link{mosaic}} which this function wraps

 \code{\link{structure}} which does the data manipulation for the crosstab
}
\keyword{crosstabs}
\keyword{mosaic}
\keyword{vcd}

