% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/max_mis.R
\name{nth_max}
\alias{nth_max}
\title{Find the nth maximum value}
\usage{
nth_max(x, n = 1)
}
\arguments{
\item{x}{a vector of numeric values}

\item{n}{which max to return}
}
\value{
the value of the nth most maximum value in a vector
}
\description{
Find the nth maximum value
}
\note{
If \code{n} is smaller/larger than \code{0}/\code{length(unique(x))}
the error \sQuote{index outside bounds} is thrown.
}
\examples{
x <- c(1:20, 20:1)
nth_max(x, n = 1) #20
nth_max(x, n = 2) #19
}

