\name{plotForWord}
\alias{plotForWord}
\title{Save a graphic as a Windows MetaFile}
\usage{
  plotForWord(x, fn)
}
\arguments{
  \item{x}{A stored plot object to be printed}

  \item{fn}{A string for the filename without any file
  extension}
}
\value{
  Nothing. Saves a .wmf with filename \code{fn} in the
  working directory.
}
\description{
  Convenience function for producing Windows MetaFile
  graphics for Microsoft documents
}

