% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/proficiency_tools.R
\name{profpoly}
\alias{profpoly}
\title{Creates a proficiency polygon in ggplot2 for showing assessment categories}
\usage{
profpoly(grades, LOSS, minimal, basic, proficient, advanced, HOSS...)
}
\arguments{
\item{grades}{a vector of tested grades in sequential order}

\item{LOSS}{is a vector of the lowest obtainable scale score on
an assessment by grade}

\item{minimal}{is a vector of the floor of the minimal assessment category by grade}

\item{basic}{is a vector of the floor of the basic assessment category by grade}

\item{proficient}{is a vector of the floor of the proficient assessment category by grade}

\item{advanced}{is a vectof of the floor of the advanced assessment category by grade}

\item{HOSS}{is a vector of the highest obtainable scale score by grade}
}
\value{
a ggplot2 object that can be printed or saved
}
\description{
Creates a proficiency polygon in ggplot2 for showing assessment categories
}
\examples{
grades<-c(3,4,5,6,7,8)
g<-length(grades)
LOSS<-rep(200,6)
HOSS<-rep(650,6)
basic<-c(320,350,370,390,420,440)
minimal<-basic-30
prof<-c(380,410,430,450,480,500)
adv<-c(480,510,530,550,580,600)

z<-profpoly(grades,LOSS,minimal,basic,proficient,advanced,HOSS)
z
}
\seealso{
\code{\link{geom_polygon}} which this function wraps
}
\keyword{ggplot2}
\keyword{polygon}

