% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dpi.R
\name{theme_dpi}
\alias{theme_dpi}
\title{a defunct ggplot2 theme developed for PDF and PNG for use at the
Wisconsin Department of Public Instruction}
\usage{
theme_dpi(base_size = 16, base_family = "")
}
\arguments{
\item{base_size}{numeric, specify the font size as a numeric value, default is 16}

\item{base_family}{character, specify the font family, this value is optional}
}
\value{
This function is defunct.
}
\description{
This was a custom ggplot2 theme developed for the Wisconsin
Department of Public Instruction. This function is now defunct.
}
\details{
All values are optional
}
\section{Deprecation}{

Defunct as of eeptools 1.3.0 (deprecated since 1.2.x). Use
\code{\link[ggplot2]{theme_bw}} instead.
}

\seealso{
\code{\link[ggplot2]{theme_bw}} from the ggplot2 package.
}
\author{
Jared E. Knowles
}
