% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dpi.R
\name{theme_dpi_map}
\alias{theme_dpi_map}
\title{a defunct ggplot2 theme developed for PDF or SVG maps}
\usage{
theme_dpi_map(base_size = 14, base_family = "")
}
\arguments{
\item{base_size}{numeric, specify the font size, default is 14}

\item{base_family}{character, specify the font family, this value is optional}
}
\value{
This function is defunct.
}
\description{
This was a ggplot2 theme developed for the Wisconsin
Department of Public Instruction for making PDF maps. This function is now defunct.
}
\section{Deprecation}{

Defunct as of eeptools 1.3.0 (deprecated since 1.2.x). Use
\code{\link[ggplot2]{theme_bw}} instead.
}

\seealso{
\code{\link[ggplot2]{theme_bw}} from the ggplot2 package.
}
\author{
Jared E. Knowles
}
