% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dpi.R
\name{theme_dpi_map2}
\alias{theme_dpi_map2}
\title{an alternate deprecated ggplot2 theme developed for PDF or SVG maps}
\source{
For more information see https://github.com/hadley/ggplot2/wiki/Themes
}
\usage{
theme_dpi_map2(base_size = 14, base_family = "")
}
\arguments{
\item{base_size}{numeric, specify the font size, default is 14}

\item{base_family}{character, specify the font family, this value is optional}
}
\value{
A theme object which is a list of attributes applied to a ggplot2 object.
}
\description{
This is a deprecated ggplot2 theme developed for the Wisconsin 
Department of Public Instruction for making PDF maps
}
\details{
All values are optional
}
\seealso{
his uses \code{\link{unit}} from the grid package extensively. 
See also \code{\link{theme_bw}} from the ggplot2 package.
}
\author{
Jared E. Knowles
}
