% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dropbox_source.R
\name{dropbox_source}
\alias{dropbox_source}
\title{A function to source R scripts from Dropbox public shares}
\usage{
dropbox_source(url)
}
\arguments{
\item{url}{a character string for the full URL of the Dropbox share}
}
\value{
None. The script is sourced.
}
\description{
Allows the user to source R scripts from a public Dropbox folder using the
URL provided by the Dropbox shared URL tool. Parses the URL and then sources
the script in one easy line.
}
\details{
url must be the full URL including https://
}
\note{
This function works with Dropbox as of the release date. If functionality
breaks, please e-mail the package maintainer to push an update.
}
\examples{
\dontrun{
 dropbox_source("https://dropbox.com/public/somekey/coolscript.R")
}
}
\author{
Jared E. Knowles
}
\seealso{
\code{\link{source}}
}
\keyword{utilities}

