% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/theme_dpi.R
\name{theme_dpi}
\alias{theme_dpi}
\title{a ggplot2 theme developed for PDF and PNG for use at the Wisconsin Department of Public Instruction}
\source{
For more information see https://github.com/hadley/ggplot2/wiki/Themes
}
\usage{
theme_dpi(base_size = 16, base_family = "")
}
\arguments{
\item{base_size}{numeric, specify the font size as a numeric value, default is 16}

\item{base_family}{character, specify the font family, this value is optional}
}
\value{
A theme object which is a list of attributes applied to a ggplot2 object.
}
\description{
This is a custom ggplot2 theme developed for the Wisconsin
Department of Public Instruction
}
\details{
All values are optional
}
\examples{
qplot(mpg, wt, data=mtcars) # standard
qplot(mpg, wt, data=mtcars) + theme_dpi()
}
\author{
Jared E. Knowles
}
\seealso{
his uses \code{\link{unit}} from the grid package extensively.
See also \code{\link{theme_bw}} from the ggplot2 package.
}

